+++ R/sha.R
@@ -16,16 +16,64 @@
 
 # nanonext - Cryptographic Hashing ---------------------------------------------
 
-#' Cryptographic Hashing Using the SHA256 Algorithm
+#' Cryptographic Hashing Using the SHA-224 Algorithm
 #'
-#' Returns a SHA256 hash or HMAC of the supplied R object. Uses the optimised
+#' Returns a SHA-224 hash or HMAC of the supplied R object. Uses the optimised
 #'     implementation from the Mbed TLS library.
 #'
 #' @param x an object.
 #' @param key (optional) supply a secret key to generate an HMAC. If missing or
-#'     NULL, the SHA256 hash of 'x' is returned.
+#'     NULL, the SHA-224 hash of 'x' is returned.
 #'
-#' @return A 'nanoHash' object - a raw vector of 32 bytes.
+#' @return A 'nanoHash' object - raw vector of length 28.
+#'
+#' @details For arguments 'x' and 'key', a raw vector is hashed directly, a
+#'     character string is converted using \code{\link{charToRaw}}, whilst other
+#'     objects are serialised first.
+#'
+#'     Use \code{as.character()} to convert the returned raw vector to a single
+#'     character string.
+#'
+#' @examples
+#' sha224("hello world!")
+#'
+#' # Converts to a character string:
+#' as.character(sha224("hello world!"))
+#'
+#' # Obtain HMAC:
+#' sha224("hello world!", "SECRET_KEY")
+#'
+#' @export
+#'
+sha224 <- function(x, key = NULL) {
+
+  if (!is.raw(x))
+    x <- if (is.character(x)) charToRaw(x) else serialize(x, NULL)
+
+  if (missing(key) || is.null(key)) {
+
+    .Call(rnng_sha224, x)
+
+  } else {
+
+    if (!is.raw(key))
+      key <- if (is.character(key)) charToRaw(key) else serialize(key, NULL)
+    .Call(rnng_sha224hmac, x, key)
+
+  }
+
+}
+
+#' Cryptographic Hashing Using the SHA-256 Algorithm
+#'
+#' Returns a SHA-256 hash or HMAC of the supplied R object. Uses the optimised
+#'     implementation from the Mbed TLS library.
+#'
+#' @param x an object.
+#' @param key (optional) supply a secret key to generate an HMAC. If missing or
+#'     NULL, the SHA-256 hash of 'x' is returned.
+#'
+#' @return A 'nanoHash' object - raw vector of length 32.
 #'
 #' @details For arguments 'x' and 'key', a raw vector is hashed directly, a
 #'     character string is converted using \code{\link{charToRaw}}, whilst other
@@ -59,6 +107,102 @@ sha256 <- function(x, key = NULL) {
     if (!is.raw(key))
       key <- if (is.character(key)) charToRaw(key) else serialize(key, NULL)
     .Call(rnng_sha256hmac, x, key)
+
+  }
+
+}
+
+#' Cryptographic Hashing Using the SHA-384 Algorithm
+#'
+#' Returns a SHA-384 hash or HMAC of the supplied R object. Uses the optimised
+#'     implementation from the Mbed TLS library.
+#'
+#' @param x an object.
+#' @param key (optional) supply a secret key to generate an HMAC. If missing or
+#'     NULL, the SHA-384 hash of 'x' is returned.
+#'
+#' @return A 'nanoHash' object - raw vector of length 48.
+#'
+#' @details For arguments 'x' and 'key', a raw vector is hashed directly, a
+#'     character string is converted using \code{\link{charToRaw}}, whilst other
+#'     objects are serialised first.
+#'
+#'     Use \code{as.character()} to convert the returned raw vector to a single
+#'     character string.
+#'
+#' @examples
+#' sha384("hello world!")
+#'
+#' # Converts to a character string:
+#' as.character(sha384("hello world!"))
+#'
+#' # Obtain HMAC:
+#' sha384("hello world!", "SECRET_KEY")
+#'
+#' @export
+#'
+sha384 <- function(x, key = NULL) {
+
+  if (!is.raw(x))
+    x <- if (is.character(x)) charToRaw(x) else serialize(x, NULL)
+
+  if (missing(key) || is.null(key)) {
+
+    .Call(rnng_sha384, x)
+
+  } else {
+
+    if (!is.raw(key))
+      key <- if (is.character(key)) charToRaw(key) else serialize(key, NULL)
+    .Call(rnng_sha384hmac, x, key)
+
+  }
+
+}
+
+#' Cryptographic Hashing Using the SHA-512 Algorithm
+#'
+#' Returns a SHA-512 hash or HMAC of the supplied R object. Uses the optimised
+#'     implementation from the Mbed TLS library.
+#'
+#' @param x an object.
+#' @param key (optional) supply a secret key to generate an HMAC. If missing or
+#'     NULL, the SHA-512 hash of 'x' is returned.
+#'
+#' @return A 'nanoHash' object - raw vector of length 64.
+#'
+#' @details For arguments 'x' and 'key', a raw vector is hashed directly, a
+#'     character string is converted using \code{\link{charToRaw}}, whilst other
+#'     objects are serialised first.
+#'
+#'     Use \code{as.character()} to convert the returned raw vector to a single
+#'     character string.
+#'
+#' @examples
+#' sha512("hello world!")
+#'
+#' # Converts to a character string:
+#' as.character(sha512("hello world!"))
+#'
+#' # Obtain HMAC:
+#' sha512("hello world!", "SECRET_KEY")
+#'
+#' @export
+#'
+sha512 <- function(x, key = NULL) {
+
+  if (!is.raw(x))
+    x <- if (is.character(x)) charToRaw(x) else serialize(x, NULL)
+
+  if (missing(key) || is.null(key)) {
+
+    .Call(rnng_sha512, x)
+
+  } else {
+
+    if (!is.raw(key))
+      key <- if (is.character(key)) charToRaw(key) else serialize(key, NULL)
+    .Call(rnng_sha512hmac, x, key)
 
   }